Use datetime format for Camdram token expiry and bigint for Camdram ids

+++ db/migrate/004_create_camdram_productions.rb
@@ -1,7 +1,7 @@
 class CreateCamdramProductions < ActiveRecord::Migration[5.2]
   def change
     create_table :camdram_productions do |t|
-      t.integer :camdram_id, null: false
+      t.bigint :camdram_id, null: false
       t.integer :max_bookings, default: 0, null: false
       t.boolean :active, default: false, null: false
       t.timestamps
